diag: cast task mask to uint
For now, only output lsw because it is too much trouble to output both.

+++ src/chain.c
@@ -114,7 +114,8 @@ void *chan_in(const char *field_name, int count, ...)
     chan_field_meta_t *latest_field = NULL;
 
     LIBCHAIN_PRINTF("[%u][0x%x & 0x%x] in: '%s':", curctx->time,
-                    curctx->self_chan_idx, curctx->task->mask, field_name);
+                    (uint16_t)curctx->self_chan_idx, (uint16_t)curctx->task->mask,
+                    field_name);
 
     va_start(ap, count);
